+++ src/main/java/com/connecter/digitalguiljabiback/controller/LoginController.java
@@ -1,10 +1,10 @@
 package com.connecter.digitalguiljabiback.controller;
 
 import com.connecter.digitalguiljabiback.domain.OauthType;
-import com.connecter.digitalguiljabiback.dto.user.LoginResponseDTO;
-import com.connecter.digitalguiljabiback.security.dto.AuthRequest;
-import com.connecter.digitalguiljabiback.security.dto.KakaoUserResponse;
-import com.connecter.digitalguiljabiback.security.dto.NaverUserResponse;
+import com.connecter.digitalguiljabiback.dto.login.LoginResponse;
+import com.connecter.digitalguiljabiback.dto.login.AuthRequest;
+import com.connecter.digitalguiljabiback.dto.login.KakaoUserResponse;
+import com.connecter.digitalguiljabiback.dto.login.NaverUserResponse;
 import com.connecter.digitalguiljabiback.security.oauth.kakao.KakaoClient;
 import com.connecter.digitalguiljabiback.security.oauth.naver.NaverClient;
 import com.connecter.digitalguiljabiback.service.LoginService;
@@ -31,22 +31,11 @@ public class LoginController {
   private final NaverClient naverClient;
   private final LoginService loginService;
 
-  /**
-   * 카카오 로그인 페이지로 리다이렉트
-   * @return ResponseEntity<Void> 리다이렉트 응답
-   */
+   //카카오 로그인 페이지 url을 반환
   @GetMapping("/kakao")
   public ResponseEntity<Map> getKakaoLoginUrl() {
     Map<String, String> map = new HashMap<>();
-    map.put("loginUrl", kakaoClient.redirectToKakaoAuth());
-    return ResponseEntity.ok(map);
-  }
-
-  @GetMapping("/naver")
-  public ResponseEntity<Map> getNaverLoginUrl() {
-    //@TODO
-    Map<String, String> map = new HashMap<>();
-    map.put("loginUrl", naverClient.redirectToNaverAuth());
+    map.put("loginUrl", kakaoClient.getAuthUrl());
     return ResponseEntity.ok(map);
   }
 
@@ -57,26 +46,43 @@ public class LoginController {
    */
   @Hidden
   @GetMapping("/callback")
-  public ResponseEntity<LoginResponseDTO> processKakaoLoginCallback(@RequestParam("code") String authorizationCode) {
+  public ResponseEntity<LoginResponse> processKakaoLoginCallback(@RequestParam("code") String authorizationCode) {
     AuthRequest params = new AuthRequest(authorizationCode);
     KakaoUserResponse response = kakaoClient.handleCallback(params);
 
     log.info("사용자 UID: {}", response.getUid());
 
-    LoginResponseDTO loginResponseDTO = loginService.loginOrCreate(response.toUserRequest(), OauthType.KAKAO);
+    LoginResponse loginResponseDTO = loginService.loginOrCreate(response.toUserRequest(), OauthType.KAKAO);
 
     return ResponseEntity.ok(loginResponseDTO);
   }
 
+  //네이버 로그인 페이지 url을 반환
+  @GetMapping("/naver")
+  public ResponseEntity<Map> getNaverLoginUrl() {
+    Map<String, String> map = new HashMap<>();
+    map.put("loginUrl", naverClient.getAuthUrl());
+    return ResponseEntity.ok(map);
+  }
+
+  /**
+   * 네이버 로그인 콜백을 처리하고 인증에 성공한 경우 로그인 또는 회원 가입
+   * @param authorizationCode 인가 코드
+   * @param state 사이트 간 요청 위조 공격을 방지하기 위해 애플리케이션에서 생성한 상태 토큰값으로 URL 인코딩을 적용한 값을 사용
+   * @return AuthResponseDTO 로그인 또는 회원 가입 결과를 담은 응답 DTO
+   */
   @Hidden
   @GetMapping("/callback/naver")
-  public ResponseEntity<LoginResponseDTO> processNaverLoginCallback(@RequestParam("code") String authorizationCode, @RequestParam("state") String state) throws UnsupportedEncodingException {
+  public ResponseEntity<LoginResponse> processNaverLoginCallback(
+    @RequestParam("code") String authorizationCode,
+    @RequestParam("state") String state
+  ) throws UnsupportedEncodingException {
     AuthRequest params = new AuthRequest(authorizationCode, state);
     NaverUserResponse response = naverClient.handleCallback(params);
 
     log.info("사용자 UID: {}", response.getId());
 
-    LoginResponseDTO loginResponseDTO = loginService.loginOrCreate(response.toUserRequest(), OauthType.NAVER);
+    LoginResponse loginResponseDTO = loginService.loginOrCreate(response.toUserRequest(), OauthType.NAVER);
 
     return ResponseEntity.ok(loginResponseDTO);
   }
