+++ src/main/java/com/connecter/digitalguiljabiback/dto/login/AuthRequest.java
@@ -1,4 +1,4 @@
-package com.connecter.digitalguiljabiback.security.dto;
+package com.connecter.digitalguiljabiback.dto.login;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,24 +7,19 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
 /**
- * KakaoAuthRequest
- * 카카오 인증 서버로 보내는 요청 객체
+ * AuthRequest
+ * 카카오, 네이버 인증 서버로 보내는 요청 객체
  * 인가 코드(authorizationCode)를 이용하여 API 요청을 생성
- *
- * 작성자: hyunjin
- * 버전: 1.0.0
- * 작성일: 2023-07-29
  */
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 public class AuthRequest {
     private String authorizationCode;
-    private String state;
+    private String state; //네이버 로그인만 사용하는 값
 
     /**
      * API 요청 바디를 생성하는 메서드
-     *
      * @return API 요청 바디
      */
     public MultiValueMap<String, String> makeBody() {
@@ -33,7 +28,7 @@ public class AuthRequest {
         body.add("state", state);
         return body;
     }
-
+    
     public AuthRequest(String authorizationCode) {
         this.authorizationCode = authorizationCode;
     }
