--- src/main/java/com/connecter/digitalguiljabiback/dto/user/LoginRequest.java
@@ -1,13 +1,0 @@
-package com.connecter.digitalguiljabiback.dto.user;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginRequest {
-  private String id;
-  private String password;
-}
