--- src/main/java/com/connecter/digitalguiljabiback/dto/user/LoginResponse.java
@@ -1,11 +1,0 @@
-package com.connecter.digitalguiljabiback.dto.user;
-
-import lombok.*;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LoginResponse {
-  private String token;
-}
