--- src/main/java/com/connecter/digitalguiljabiback/dto/user/RegisterRequest.java
@@ -1,13 +1,0 @@
-package com.connecter.digitalguiljabiback.dto.user;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequest {
-  private String id;
-  private String password;
-}
