--- src/main/java/com/connecter/digitalguiljabiback/dto/user/RegisterResponse.java
@@ -1,12 +1,0 @@
-package com.connecter.digitalguiljabiback.dto.user;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-@AllArgsConstructor
-public class RegisterResponse {
-  private String id;
-}
