--- src/main/java/com/connecter/digitalguiljabiback/security/jwt/JwtProvider.java
@@ -1,43 +1,0 @@
-package com.connecter.digitalguiljabiback.security.jwt;
-
-import com.connecter.digitalguiljabiback.config.properties.JwtProperties;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.spec.SecretKeySpec;
-import java.sql.Timestamp;
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
-import java.util.Date;
-
-@Component
-@RequiredArgsConstructor
-public class JwtProvider {
-    private final JwtProperties jwtProperties;
-
-    public String createToken(String userSpecification) {
-        String secretKey = jwtProperties.getSecretKey();
-        long expirationHours = jwtProperties.getExpirationHours();
-        String issuer = jwtProperties.getIssuer();
-
-        return Jwts.builder()
-                .signWith(SignatureAlgorithm.HS512, new SecretKeySpec(secretKey.getBytes(), SignatureAlgorithm.HS512.getJcaName()))
-                .setSubject(userSpecification)
-                .setIssuer(issuer)
-                .setIssuedAt(Timestamp.valueOf(LocalDateTime.now()))
-                .setExpiration(Date.from(Instant.now().plus(expirationHours, ChronoUnit.HOURS)))
-                .compact();
-    }
-
-    public String validateTokenAndGetSubject(String token) {
-        String secretKey = jwtProperties.getSecretKey();
-        return Jwts.parser()
-                .setSigningKey(secretKey.getBytes())
-                .parseClaimsJws(token)
-                .getBody()
-                .getSubject();
-    }
-}
